Move git operations to it's own class
