+++ lib/creds.js
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import path from 'path';
 import AWS from 'aws-sdk';
-import cp from 'child_process';
+import Git from './git.js';
 import AJV from 'ajv';
 
 const ajv = new AJV({
@@ -45,11 +45,11 @@ const PROFILE_KEYS = [
  */
 export default class Credentials {
     constructor (argv, opts) {
-        this.user = Credentials.gituser();
-        this.owner = Credentials.gitowner();
+        this.user = Git.user();
+        this.owner = Git.owner();
 
-        this.repo = Credentials.gitrepo();
-        this.sha = Credentials.gitsha();
+        this.repo = Git.repo();
+        this.sha = Git.sha();
 
         // If GH deployment has been requested, store ID here
         this.deployment = false;
@@ -185,11 +185,13 @@ export default class Credentials {
 
     /**
      * Attempt to read a dot deploy file
+     *
+     * @returns {Object} Dot Deploy Object
      */
     static dot_deploy() {
         const attempts = [
             path.resolve('./.deploy'),
-            path.resolve(Credentials.gitroot(), '.deploy')
+            path.resolve(Git.root(), '.deploy')
         ];
 
         let dotdeploy = false;
@@ -215,83 +217,6 @@ export default class Credentials {
         }
 
         return dotdeploy;
-    }
-
-    /**
-     * Return top level dir of a git repo
-     * @return {string}
-     */
-    static gitroot() {
-        const git = cp.spawnSync('git', [
-            'rev-parse', '--show-toplevel'
-        ]);
-
-        if (!git.stdout) return (new Error('Is this a git repo? Could not determine Git Root Directory'));
-        return String(git.stdout).replace(/\n/g, '');
-    }
-
-    /**
-     * Get the name of the current GitRepo
-     * @return {string}
-     */
-    static gitrepo() {
-        return path.parse(Credentials.gitroot()).name;
-    }
-
-    /**
-     * Get the name of the current git user
-     * @return {string}
-     */
-    static gituser() {
-        const git = cp.spawnSync('git', [
-            'config', 'user.name'
-        ]);
-
-        if (!git.stdout) return (new Error('Is this a git repo? Could not determine GitSha'));
-        return String(git.stdout).replace(/\n/g, '');
-    }
-
-    /**
-     * Get the name of the upstream git owner
-     * @return {string}
-     */
-    static gitowner() {
-        const git = cp.spawnSync('git', [
-            'config', '--get', 'remote.origin.url'
-        ]);
-
-        if (!String(git.stdout)) return false;
-
-        const owner = String(git.stdout).replace(/\n/g, '');
-
-        if (owner.includes('git@github.com')) {
-            return owner
-                .replace(/.*git@github.com:/, '')
-                .replace(/\/.*/, '');
-        } else if (owner.includes('https://github.com')) {
-            const giturl = new URL(owner);
-
-            return giturl.pathname
-                .replace('.git', '')
-                .slice(1)
-                .replace(/\/.*/, '');
-        } else {
-            throw new Error('only origins of format: git@github.com or https://github.com are supported');
-        }
-    }
-
-    /**
-     * Get the current GitSha
-     */
-    static gitsha() {
-        const git = cp.spawnSync('git', [
-            '--git-dir', path.resolve(Credentials.gitroot(), '.git'),
-            'rev-parse', 'HEAD'
-        ]);
-
-        if (!git.stdout) return (new Error('Is this a git repo? Could not determine GitSha'));
-        return String(git.stdout).replace(/\n/g, '');
-
     }
 
     async accountId() {
