+++ lib/list.js
@@ -22,6 +22,9 @@ export default class List {
 
     /**
      * List current stacks deployed to a given profile
+     *
+     * @param {Credentials} creds
+     * @param {Object} argv
      */
     static async main(creds, argv) {
         argv = minimist(argv, {
